requirements.txt
transformers==4.18.0
sentencepiece==0.1.96
pytorch_lightning==0.8.1
editdistance==0.6.0
scikit-learn==0.24.2
numpy==1.22.3
tqdm==4.64.0
